+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Use websocket network transport by default
 - Add option Settings->Server->Use websocket transport
 - Add ig.ENTITY.WaterBubblePanel state
+- Add remote combatant elemental inflict status bar state
+- Add remote combatant break bar state
 
 ### Fixed